BPMN added

+++ PapaCarloDBApp/PapaCarloDBApp/QueryReports.cs
@@ -9,103 +9,123 @@ namespace PapaCarloDBApp
     {
         public List<ReportBalanceTable> querySelectReportBalance(DateTime date)
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                var query = from c in db.ReportBalances
-                            join sc in db.StoreCells on c.StoreCellId equals sc.Id
-                            join sh in db.Storehouses on sc.StorehouseId equals sh.Id
-                            join p in db.Products on c.ProductId equals p.Id
-                            where c.Date == date
-                            select new { c, p, sc, sh };
-
-                List<ReportBalanceTable> table = new List<ReportBalanceTable>();
-
-                foreach (var item in query)
+                using (DataBaseContext db = new DataBaseContext())
                 {
-                    table.Add(new ReportBalanceTable(item.c, item.p, item.sc, item.sh));
+                    var query = from c in db.ReportBalances
+                                join sc in db.StoreCells on c.StoreCellId equals sc.Id
+                                join sh in db.Storehouses on sc.StorehouseId equals sh.Id
+                                join p in db.Products on c.ProductId equals p.Id
+                                where c.Date == date
+                                select new { c, p, sc, sh };
+
+                    List<ReportBalanceTable> table = new List<ReportBalanceTable>();
+
+                    foreach (var item in query)
+                    {
+                        table.Add(new ReportBalanceTable(item.c, item.p, item.sc, item.sh));
+                    }
+                    return table;
                 }
-                return table;
             }
+            return null;
         }
 
         public List<ReportBalanceTable> querySelectReportBalance()
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                var query = from c in db.ReportBalances
-                            join sc in db.StoreCells on c.StoreCellId equals sc.Id
-                            join sh in db.Storehouses on sc.StorehouseId equals sh.Id
-                            join p in db.Products on c.ProductId equals p.Id
-                            select new { c, p, sc, sh };
-
-                List<ReportBalanceTable> table = new List<ReportBalanceTable>();
-
-                foreach (var item in query)
+                using (DataBaseContext db = new DataBaseContext())
                 {
-                    table.Add(new ReportBalanceTable(item.c, item.p, item.sc, item.sh));
+                    var query = from c in db.ReportBalances
+                                join sc in db.StoreCells on c.StoreCellId equals sc.Id
+                                join sh in db.Storehouses on sc.StorehouseId equals sh.Id
+                                join p in db.Products on c.ProductId equals p.Id
+                                select new { c, p, sc, sh };
+
+                    List<ReportBalanceTable> table = new List<ReportBalanceTable>();
+
+                    foreach (var item in query)
+                    {
+                        table.Add(new ReportBalanceTable(item.c, item.p, item.sc, item.sh));
+                    }
+                    return table;
                 }
-                return table;
             }
+            return null;
         }
 
         public List<ReportPredictTable> querySelectReportPredict(DateTime date)
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                var query = from c in db.ReportPredicts
-                            join sh in db.Storehouses on c.StorehouseId equals sh.Id
-                            join p in db.Products on c.ProductId equals p.Id
-                            //join cs in db.ContractSupplys on c.ContractSupplyId equals cs.Id
-                            where c.Date == date
-                            select new { c, p, sh };
-
-                List<ReportPredictTable> table = new List<ReportPredictTable>();
-
-                foreach (var item in query)
+                using (DataBaseContext db = new DataBaseContext())
                 {
-                    table.Add(new ReportPredictTable(item.c, item.p, item.sh));
+                    var query = from c in db.ReportPredicts
+                                join sh in db.Storehouses on c.StorehouseId equals sh.Id
+                                join p in db.Products on c.ProductId equals p.Id
+                                //join cs in db.ContractSupplys on c.ContractSupplyId equals cs.Id
+                                where c.Date == date
+                                select new { c, p, sh };
+
+                    List<ReportPredictTable> table = new List<ReportPredictTable>();
+
+                    foreach (var item in query)
+                    {
+                        table.Add(new ReportPredictTable(item.c, item.p, item.sh));
+                    }
+                    return table;
                 }
-                return table;
             }
+            return null;
         }
 
         public List<ReportPredictTable> querySelectReportPredict()
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                var query = from c in db.ReportPredicts
-                            join sh in db.Storehouses on c.StorehouseId equals sh.Id
-                            join p in db.Products on c.ProductId equals p.Id
-                            select new { c, p, sh };
-
-                List<ReportPredictTable> table = new List<ReportPredictTable>();
-
-                foreach (var item in query)
+                using (DataBaseContext db = new DataBaseContext())
                 {
-                    table.Add(new ReportPredictTable(item.c, item.p, item.sh));
+                    var query = from c in db.ReportPredicts
+                                join sh in db.Storehouses on c.StorehouseId equals sh.Id
+                                join p in db.Products on c.ProductId equals p.Id
+                                select new { c, p, sh };
+
+                    List<ReportPredictTable> table = new List<ReportPredictTable>();
+
+                    foreach (var item in query)
+                    {
+                        table.Add(new ReportPredictTable(item.c, item.p, item.sh));
+                    }
+                    return table;
                 }
-                return table;
             }
+            return null;
         }
 
         public bool queryAddReportBalance(ReportBalance c)
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                try
+                using (DataBaseContext db = new DataBaseContext())
                 {
-                    db.ReportBalances.Add(c);
-                    db.SaveChanges();
+                    try
+                    {
+                        db.ReportBalances.Add(c);
+                        db.SaveChanges();
+                    }
+                    catch (Exception e)
+                    {
+                        return false;
+                    }
+                    return true;
                 }
-                catch (Exception e)
-                {
-                    return false;
-                }
-                return true;
             }
+            return false;
         }
 
-        public bool queryUpdateReportBalance(ReportBalance c)
+        /*public bool queryUpdateReportBalance(ReportBalance c)
         {
             using (DataBaseContext db = new DataBaseContext())
             {
@@ -126,53 +146,65 @@ namespace PapaCarloDBApp
                 }
                 return true;
             }
-        }
+        }*/
 
         public bool queryDeleteReportBalance(int Id)
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                try
+                using (DataBaseContext db = new DataBaseContext())
                 {
-                    ReportBalance p = db.ReportBalances.Find(Id);
-                    db.ReportBalances.Remove(p);
+                    try
+                    {
+                        ReportBalance p = db.ReportBalances.Find(Id);
+                        db.ReportBalances.Remove(p);
 
-                    db.SaveChanges();
+                        db.SaveChanges();
+                    }
+                    catch (Exception e)
+                    {
+                        return false;
+                    }
+                    return true;
                 }
-                catch (Exception e)
-                {
-                    return false;
-                }
-                return true;
             }
+            return false;
         }
 
         public ReportBalance queryFindReportBalanceById(int Id)
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                return db.ReportBalances.Find(Id);
+                using (DataBaseContext db = new DataBaseContext())
+                {
+                    return db.ReportBalances.Find(Id);
+                }
             }
+            return null;
         }
 
         public bool queryAddReportPredict(ReportPredict c)
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                try
+                using (DataBaseContext db = new DataBaseContext())
                 {
-                    db.ReportPredicts.Add(c);
-                    db.SaveChanges();
+                    try
+                    {
+                        db.ReportPredicts.Add(c);
+                        db.SaveChanges();
+                    }
+                    catch (Exception e)
+                    {
+                        return false;
+                    }
+                    return true;
                 }
-                catch (Exception e)
-                {
-                    return false;
-                }
-                return true;
             }
+            return false;
         }
 
-        public bool queryUpdateReportPredict(ReportPredict c)
+       /* public bool queryUpdateReportPredict(ReportPredict c)
         {
             using (DataBaseContext db = new DataBaseContext())
             {
@@ -195,33 +227,41 @@ namespace PapaCarloDBApp
                 }
                 return true;
             }
-        }
+        }*/
 
         public bool queryDeleteReportPredict(int Id)
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                try
+                using (DataBaseContext db = new DataBaseContext())
                 {
-                    ReportPredict p = db.ReportPredicts.Find(Id);
-                    db.ReportPredicts.Remove(p);
+                    try
+                    {
+                        ReportPredict p = db.ReportPredicts.Find(Id);
+                        db.ReportPredicts.Remove(p);
 
-                    db.SaveChanges();
+                        db.SaveChanges();
+                    }
+                    catch (Exception e)
+                    {
+                        return false;
+                    }
+                    return true;
                 }
-                catch (Exception e)
-                {
-                    return false;
-                }
-                return true;
             }
+            return false;
         }
 
         public ReportPredict queryFindReportPredictById(int Id)
         {
-            using (DataBaseContext db = new DataBaseContext())
+            if (LoginInfo.Position == 1 || LoginInfo.Position == 2)//Начальник склада, Бухгалтер  
             {
-                return db.ReportPredicts.Find(Id);
+                using (DataBaseContext db = new DataBaseContext())
+                {
+                    return db.ReportPredicts.Find(Id);
+                }
             }
+            return null;
         }
     }
 }
